requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector==2.2.9
mysql-connector-python==9.3.0
Werkzeug==3.1.3
pip install -r requirements.txt     #just to restart
1. python -m venv venv
2. .\venv\Scripts\activate
3. pip install flask mysql-connector-python
4. python app.py
